# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GitLab User Reports",
        "version": "0.1.0"
    },
    "paths": {
...
        "/projects/{project_id}": {
            "get": {
                "tags": ["projects"],
                "summary": "Get Project",
                "description": "Get details of a specific GitLab project.",
                "operationId": "get_project_projects__project_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "GitLab project ID",
                            "title": "Project Id"
                        },
                        "description": "GitLab project ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{user_id}": {
            "get": {
                "tags": ["users"],
                "summary": "Get Gitlab User",
                "description": "Retrieve a GitLab user by their ID.",
                "operationId": "get_gitlab_user_users__user_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "The ID of the GitLab user to retrieve.",
                            "title": "User Id"
                        },
                        "description": "The ID of the GitLab user to retrieve."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GitLabUser"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/performance/users/{user_id}": {
            "get": {
                "tags": ["performance"],
                "summary": "Get User Performance",
                "description": "Calculate and retrieve user performance data over a specified time period.",
                "operationId": "get_user_performance_performance_users__user_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "The ID of the GitLab user.",
                            "title": "User Id"
                        },
                        "description": "The ID of the GitLab user."
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "description": "The start date for the performance period in ISO format.",
                            "title": "Start Date"
                        },
                        "description": "The start date for the performance period in ISO format."
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "description": "The end date for the performance period in ISO format.",
                            "title": "End Date"
                        },
                        "description": "The end date for the performance period in ISO format."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralUserPerformance"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/performance/projects/{project_id}/users/{user_id}": {
            "get": {
                "tags": ["performance"],
                "summary": "Get Project Performance",
                "description": "Calculate and retrieve project performance data over a specified time period.",
                "operationId": "get_project_performance_performance_projects__project_id__users__user_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "The ID of the GitLab project.",
                            "title": "Project Id"
                        },
                        "description": "The ID of the GitLab project."
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "The ID of the GitLab user.",
                            "title": "User Id"
                        },
                        "description": "The ID of the GitLab user."
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "description": "The start date for the performance period in ISO format.",
                            "title": "Start Date"
                        },
                        "description": "The start date for the performance period in ISO format."
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "description": "The end date for the performance period in ISO format.",
                            "title": "End Date"
                        },
                        "description": "The end date for the performance period in ISO format."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectPerformanceResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/performance/projects/{project_id}": {
            "get": {
                "tags": ["performance"],
                "summary": "Get General Project Performance",
                "description": "Calculate and retrieve general project performance data over a specified time period.",
                "operationId": "get_general_project_performance_performance_projects__project_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "The ID of the GitLab project.",
                            "title": "Project Id"
                        },
                        "description": "The ID of the GitLab project."
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "description": "The start date for the performance period in ISO format.",
                            "title": "Start Date"
                        },
                        "description": "The start date for the performance period in ISO format."
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "description": "The end date for the performance period in ISO format.",
                            "title": "End Date"
                        },
                        "description": "The end date for the performance period in ISO format."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralProjectPerformance"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponseModel": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": ["detail"],
                "title": "ErrorResponseModel",
                "description": "Standard error response model."
            },
...
            "ProjectsResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id",
                        "description": "Project ID"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "Project name"
                    },
                    "name_with_namespace": {
                        "type": "string",
                        "title": "Name With Namespace",
                        "description": "Project name with namespace"
                    },
                    "path_with_namespace": {
                        "type": "string",
                        "title": "Path With Namespace",
                        "description": "Project path with namespace"
                    },
                    "tag_list": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tag List",
                        "description": "List of project tags"
                    },
                    "topics": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Topics",
                        "description": "List of project topics"
                    },
                    "web_url": {
                        "type": "string",
                        "title": "Web Url",
                        "description": "Web URL to the project"
                    },
                    "avatar_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Avatar Url",
                        "description": "URL to the project's avatar"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At",
                        "description": "Project creation timestamp"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "name_with_namespace",
                    "path_with_namespace",
                    "tag_list",
                    "topics",
                    "web_url",
                    "created_at"
                ],
                "title": "ProjectsResponse",
                "description": "Response schema for GitLab projects."
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}